+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.2.0] - 2021-07-07
 ### Changed
 - Move Colors class to colors.py.
+### Fix
+- Incorrect type of tags in tasks.py.
 
 ## [0.1.0] - 2021-07-06
 ### Added